add photo management

+++ src/main/java/ru/hits/just_4sport/service/UserProfileService.java
@@ -3,11 +3,13 @@ package ru.hits.just_4sport.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 import ru.hits.just_4sport.infrastructure.exception.NotFoundException;
-import ru.hits.just_4sport.model.api.user.UserPhotoModel;
 import ru.hits.just_4sport.model.api.user.UserProfileModel;
 import ru.hits.just_4sport.model.api.user.UserUpdateProfileModel;
 import ru.hits.just_4sport.model.domain.PhotoEntity;
+import ru.hits.just_4sport.model.mapper.PhotoMapper;
+import ru.hits.just_4sport.properties.UploadProperties;
 import ru.hits.just_4sport.repository.EventRepository;
 import ru.hits.just_4sport.repository.PhotoRepository;
 import ru.hits.just_4sport.repository.RefreshTokenRepository;
@@ -23,6 +25,9 @@ public class UserProfileService {
     private final EventRepository eventsRepository;
     private final PhotoRepository photoRepository;
     private final RefreshTokenRepository refreshTokenRepository;
+    private final PhotoMapper photoMapper;
+    private final PhotoService photoService;
+    private final UploadProperties uploadProperties;
 
     public UserProfileModel getUserProfile(UUID id) {
         var user = userRepository.findById(id)
@@ -32,7 +37,7 @@ public class UserProfileService {
                 .setName(user.getName())
                 .setNickname(user.getNickname())
                 .setEmail(user.getEmail())
-                .setPhoto(user.getPhoto())
+                .setPhoto(photoMapper.toModel(user.getPhoto()))
                 .setFavoriteSports(user.getFavoriteSports())
                 .setAuthorEvents(user.getAuthorEvents());
 
@@ -69,31 +74,43 @@ public class UserProfileService {
         userRepository.deleteById(id);
     }
 
-    public void setUserProfilePhoto(UUID id, UserPhotoModel photo) {
-        var user = userRepository.findById(id)
+    public void setUserProfilePhoto(String email, MultipartFile photo) {
+        var user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
 
-        var newPhoto = new PhotoEntity()
-                .setTitle(photo.getTitle())
-                .setPath(photo.getPath());
+        photoService.validateImage(photo);
 
-        photoRepository.save(newPhoto);
+        String photoName = photoService.saveImage(photo);
 
-        user.setPhoto(newPhoto);
+        var photoEntity = new PhotoEntity()
+                .setTitle(photo.getOriginalFilename())
+                .setPath(uploadProperties.profilePhotosDir() + photoName);
 
+        var oldPhoto = user.getPhoto();
+
+        if (oldPhoto != null) {
+            photoService.deleteImage(oldPhoto.getPath());
+        }
+
+        user.setPhoto(photoEntity);
         userRepository.save(user);
     }
 
-    public void deleteUserProfilePhoto(UUID id) {
-        var user = userRepository.findById(id)
+    public void deleteUserProfilePhoto(String email) {
+        var user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
 
         var photo = user.getPhoto();
 
+        if (photo ==  null) {
+            return;
+        }
+
         photoRepository.delete(photo);
 
         user.setPhoto(null);
-
         userRepository.save(user);
+
+        photoService.deleteImage(photo.getPath());
     }
 }
